The September Update: - Fixed Matrix4 rotation bug - Removed Windows dependencies - Added Linux platform code for console utilities (needs testing) - Improved debugging - Now OBBs don't use Matrix4 for rotation - Removed broken OBB code - Optimized OBBs - Improved Shaders - Added more platform macros - Changed DELTAENGINE_MSG with DELTAENGINE_DBG - Fixed compilation warnings - Added memory manager macro definitions - Now Debug::breakpoint can show the file and line of the caller - Now init prints the build time - Added functions to print dates, time and timestamps - FileIO & Logger now log to the console on error - Removed errors.h

+++ DeltaEngine/src/graphics/Window.cpp
@@ -94,7 +94,7 @@ namespace DeltaEngine {
 
 			modifier &= 0xFC;
 
-			window->keys[key] = (repeat << 1) | pressed;
+			window->keys[key] = ((int)repeat << 1) | (int)pressed;
 
 			//TODO: Check this!
  			if (pressed)
